requirements.txt
Django==2.2
djangorestframework==3.9.2
Pillow==6.0.0
pytz==2019.1
sqlparse==0.3.0
django-filter==2.1.0
psycopg2==2.8.2

